refactor(config): extract identity helpers to stay under 300-line limit
